fix: External links and multiple initializations of Collapsable
When `Collapsable` is initialized more than once on a page and there is some external link, there is an error thrown into the console `Error: Collapsable: External link has no associated collapsable item.` as a result of some instances not being able to recognize the external link. From now, this is silently ignored as not an error. Also extLink