feature: Allow external links to be a button element
External links for collapsable can now be a `HTMLButtonElement` with `data-collapsable-id` attribute. There is also an option to prevent the default action of an external link using either the option (this affects all external links for a particular `Collapsable`) or the `data-collapsable-prevent-default` attribute (this affects only a specific external link).